Organize JS files
