+++ index.html
@@ -132,6 +132,6 @@
         <p>Select a markdown file from the list to start editing, or create a new file.</p>
     </div>
     
-    <script src="app.js"></script>
+    <script src="js/plugin.js"></script>
 </body>
 </html>
